requirements.txt
# Deep Learning Framework
torch>=1.10.0
torchvision>=0.11.0

# Web Interface
gradio>=4.0.0

# Data Processing
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0
openpyxl>=3.0.0

# Model Analysis
thop>=0.1.1
pycm>=3.5.0
prettytable>=2.0.0
pytorch-grad-cam>=1.4.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Logging
wandb>=0.13.0

# Archive Handling
patoolib>=1.12.0

# LLM API (Optional - requires DASHSCOPE_API_KEY environment variable)
dashscope>=1.10.0

# Other Utilities
tqdm>=4.62.0
Pillow>=8.0.0
opencv-python>=4.5.0
joblib>=1.1.0
